withplay/yrequirement.txt
django==2.1.7
djangorestframework
django-rest-swagger
django-redis
pymysql
django-filter
demjson
pillow
redis==2.10.6
